% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue.viz.table.fg_matrix}
\alias{catalogue.viz.table.fg_matrix}
\title{Internal function computing the foreground color matrix for the catalogue}
\usage{
catalogue.viz.table.fg_matrix(data_matrix)
}
\arguments{
\item{data_matrix}{the catalogue data matrix}
}
\value{
a table with the specifications of the foreground color of each data cell in the catalogue
}
\description{
Internal function computing the foreground color matrix for the catalogue
}
