% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue.viz.table.legend.colours}
\alias{catalogue.viz.table.legend.colours}
\title{Produces a _flextable_ corresponding to the color legend used for the catalogue cell content}
\usage{
catalogue.viz.table.legend.colours()
}
\value{
a _flextable_ corresponding to the lcolor legend used for the catalogue cell content
}
\description{
Produces a _flextable_ corresponding to the color legend used for the catalogue cell content
}
