% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue.viz.table.legend}
\alias{catalogue.viz.table.legend}
\title{Produces a _flextable_ corresponding to the legend of the catalogue cell content}
\usage{
catalogue.viz.table.legend()
}
\value{
a _flextable_ corresponding to the legend of the catalogue cell content
}
\description{
Produces a _flextable_ corresponding to the legend of the catalogue cell content
}
