% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue.viz.table}
\alias{catalogue.viz.table}
\title{TBD}
\usage{
catalogue.viz.table(
  catalogue_data,
  show_catches_gradient = TRUE,
  remove_species = FALSE,
  remove_stock = FALSE,
  truncate_years = TRUE
)
}
\arguments{
\item{catalogue_data}{TBD}

\item{remove_species}{TBD}

\item{remove_stock}{TBD}

\item{truncate_years}{TBD}
}
\value{
TBD
}
\description{
TBD
}
