% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catdis_maps.R
\name{catdis.plot.heatmap}
\alias{catdis.plot.heatmap}
\title{TBD}
\usage{
catdis.plot.heatmap(
  base_map = map.atlantic(),
  catdis_data,
  gears_to_keep = NULL,
  gear,
  num_breaks = 5
)
}
\arguments{
\item{base_map}{TBD}

\item{catdis_data}{TBD}

\item{gears_to_keep}{TBD}

\item{gear}{TBD}

\item{num_breaks}{TBD}
}
\value{
TBD
}
\description{
TBD
}
