% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.viz.executive_summary.table.all.species_group.xlsx}
\alias{t1nc.viz.executive_summary.table.all.species_group.xlsx}
\title{Produces a standalone Excel file containing the global table of T1 nominal catch data summary **for a group of species** as expected in the SCRS executive summary tables.
This version}
\usage{
t1nc.viz.executive_summary.table.all.species_group.xlsx(
  filtered_t1nc_data,
  species_group_code,
  species_group_descriptions,
  output_file,
  version = 0,
  fill = NA,
  legacy_style = FALSE,
  show_grids = FALSE
)
}
\arguments{
\item{species_group_code}{a species group code}

\item{species_group_descriptions}{a list containing the species group scientific name as well as its name in the three official ICCAT languages}

\item{output_file}{the output Excel file name}

\item{version}{to specify the table version in accordance with the SCRS executive summary tables template}

\item{fill}{the value to use to fill cells with no catch data}

\item{legacy_style}{whether or not applying the 'legacy' table output style}

\item{show_grids}{whether or not showing cell grids in the Excel spreadsheet}

\item{t1nc_data}{T1 nominal catch data as retrieved using the \code{\link{iccat.dev.data::t1nc}} function}
}
\value{
nothing, as the Excel file is simply stored at the provided path
}
\description{
Produces a standalone Excel file containing the global table of T1 nominal catch data summary **for a group of species** as expected in the SCRS executive summary tables.
This version
}
