% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.viz.executive_summary.table.gears}
\alias{t1nc.viz.executive_summary.table.gears}
\title{TBD}
\usage{
t1nc.viz.executive_summary.table.gears(t1nc_data)
}
\arguments{
\item{t1nc_data}{TBD}
}
\value{
TBD
}
\description{
TBD
}
