% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.viz.trends.legend}
\alias{t1nc.viz.trends.legend}
\title{A printable version of the T1NC trends legend.}
\usage{
t1nc.viz.trends.legend()
}
\value{
a _flextable_ formatted according to the agreed style guides
}
\description{
A printable version of the T1NC trends legend.
}
