% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.viz.trends}
\alias{t1nc.viz.trends}
\title{TBD}
\usage{
t1nc.viz.trends(
  t1nc_data,
  year_min = NA,
  year_max = NA,
  by_species = TRUE,
  by_stock = TRUE,
  by_gear = TRUE,
  rel_diff_limits = DEFAULT_TRENDS_REL_DIFF_LIMITS,
  sensitivity = 0,
  colorize_gears = FALSE
)
}
\arguments{
\item{t1nc_data}{TBD}

\item{by_species}{TBD}

\item{by_stock}{TBD}

\item{by_gear}{TBD}

\item{rel_diff_limits}{TBD}

\item{sensitivity}{TBD}

\item{colorize_gears}{TBD}
}
\value{
TBD
}
\description{
TBD
}
