% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue.viz.table.fg_matrix}
\alias{catalogue.viz.table.fg_matrix}
\title{TBD}
\usage{
catalogue.viz.table.fg_matrix(data_matrix)
}
\arguments{
\item{data_matrix}{TBD}
}
\value{
TBD
}
\description{
TBD
}
