% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue.viz.table}
\alias{catalogue.viz.table}
\title{Produces a _flextable_ containing a formatted SCRS catalogue}
\usage{
catalogue.viz.table(
  catalogue_data,
  show_catches_gradient = TRUE,
  remove_species = FALSE,
  remove_stock = FALSE,
  truncate_years = TRUE,
  flag_separator_width = 1,
  default_font_size = 7,
  values_font_size = 7,
  default_h_padding = 5,
  values_h_padding = 2
)
}
\arguments{
\item{catalogue_data}{the SCRS catalogue data as produced using \code{\link{iccat.pub.data::catalogue.compile}}, with T1 nominal catch data and base catalogue data extracted from the ICCAT databases through the **iccat.dev.data** and **iccat.pub.data** libraries}

\item{show_catches_gradient}{to enable / disable a color gradients on the cumulative catch column}

\item{remove_species}{if the species shall not be included in the stratification}

\item{remove_stock}{if the stock shall not be included in the stratification}

\item{truncate_years}{if year values in the table headers should be truncated to the last two digits}

\item{flag_separator_width}{TBD}

\item{default_font_size}{the font size applied by default}

\item{values_font_size}{the font size used for value cells only}

\item{default_h_padding}{the horizontal padding applied to all cells by default}

\item{values_h_padding}{the horizontal padding appliet to value cells only}
}
\value{
a _flextable_ containing a SCRS catalogue formatted according to the provided criteria
}
\description{
Produces a _flextable_ containing a formatted SCRS catalogue
}
