% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue.viz.table}
\alias{catalogue.viz.table}
\title{TBD}
\usage{
catalogue.viz.table(
  catalogue_data,
  show_catches_gradient = TRUE,
  remove_species = FALSE,
  remove_stock = FALSE,
  truncate_years = TRUE,
  flag_separator_width = 1,
  default_font_size = 7,
  values_font_size = 7,
  default_h_padding = 5,
  values_h_padding = 2
)
}
\arguments{
\item{catalogue_data}{TBD}

\item{remove_species}{TBD}

\item{remove_stock}{TBD}

\item{truncate_years}{TBD}

\item{flag_separator_width}{TBD}

\item{default_font_size}{TBD}

\item{values_font_size}{TBD}

\item{default_h_padding}{TBD}

\item{values_h_padding}{TBD}
}
\value{
TBD
}
\description{
TBD
}
