% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue.viz.table.xlsx.append}
\alias{catalogue.viz.table.xlsx.append}
\title{TBD}
\usage{
catalogue.viz.table.xlsx.append(
  filtered_catalogue_data,
  workbook,
  pretty_print_catches = FALSE,
  cutoff_percentage = 95,
  max_percentage = 100,
  score,
  table_number,
  stock
)
}
\arguments{
\item{filtered_catalogue_data}{TBD}

\item{workbook}{TBD}

\item{pretty_print_catches}{TBD}

\item{cutoff_percentage}{TBD}

\item{max_percentage}{TBD}

\item{score}{TBD}

\item{table_number}{TBD}

\item{stock}{TBD}
}
\value{
TBD
}
\description{
TBD
}
