% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\docType{data}
\name{DEFAULT_TRENDS_REL_DIFF_LIMITS}
\alias{DEFAULT_TRENDS_REL_DIFF_LIMITS}
\title{The table modeling the default T1NC trends classification limits with respect to the
relative magnitude of catch differences.}
\format{
An object of class \code{list} of length 4.
}
\usage{
DEFAULT_TRENDS_REL_DIFF_LIMITS
}
\description{
The table modeling the default T1NC trends classification limits with respect to the
relative magnitude of catch differences.
}
\keyword{datasets}
