% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.viz.executive_summary.table.all.species_group.xlsx}
\alias{t1nc.viz.executive_summary.table.all.species_group.xlsx}
\title{TBD}
\usage{
t1nc.viz.executive_summary.table.all.species_group.xlsx(
  filtered_t1nc_data,
  species_group_code,
  species_group_descriptions,
  output_file,
  version = 0,
  fill = NA
)
}
\arguments{
\item{species_group_code}{TBD}

\item{species_group_descriptions}{TBD}

\item{output_file}{TBD}

\item{version}{TBD}

\item{fill}{TBD}

\item{t1nc_data}{TBD}
}
\value{
TBD
}
\description{
TBD
}
