% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.viz.executive_summary.table.all.xlsx}
\alias{t1nc.viz.executive_summary.table.all.xlsx}
\title{TBD}
\usage{
t1nc.viz.executive_summary.table.all.xlsx(t1nc_data, output_file)
}
\arguments{
\item{t1nc_data}{TBD}

\item{output_file}{TBD}
}
\value{
TBD
}
\description{
TBD
}
