% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.viz.executive_summary.table.all.xlsx}
\alias{t1nc.viz.executive_summary.table.all.xlsx}
\title{TBD}
\usage{
t1nc.viz.executive_summary.table.all.xlsx(
  t1nc_data,
  output_file,
  version = 0,
  fill = NA
)
}
\arguments{
\item{t1nc_data}{TBD}

\item{output_file}{TBD}

\item{version}{TBD}

\item{fill}{TBD}
}
\value{
TBD
}
\description{
TBD
}
