% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.viz.executive_summary.table.CPCs}
\alias{t1nc.viz.executive_summary.table.CPCs}
\title{Produces a table of T1 nominal catch data summary as expected in the _CPC_ section of the SCRS executive summary tables}
\usage{
t1nc.viz.executive_summary.table.CPCs(t1nc_data, fill = NA)
}
\arguments{
\item{t1nc_data}{T1 nominal catch data as retrieved using the \code{\link{iccat.dev.data::t1nc}} function}

\item{fill}{the value to use to fill cells with no catch data}
}
\value{
a _flextable_ containing the information to be included in the _CPC_ section of the SCRS executive summaries
}
\description{
Produces a table of T1 nominal catch data summary as expected in the _CPC_ section of the SCRS executive summary tables
}
