% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.viz.executive_summary.table.global}
\alias{t1nc.viz.executive_summary.table.global}
\title{TBD}
\usage{
t1nc.viz.executive_summary.table.global(t1nc_data, fill = NA)
}
\arguments{
\item{t1nc_data}{TBD}

\item{fill}{TBD}
}
\value{
TBD
}
\description{
TBD
}
